Backend/requirements.txt
fastapi
firebase-admin
python-dotenv
pydantic
twilio
uvicorn
